refactor: split copilot local session flow
