+++ internal/providers/copilot/copilot.go
@@ -7,14 +7,11 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"sort"
 	"strings"
 	"time"
 
 	"github.com/janekbaraniewski/openusage/internal/core"
 	"github.com/janekbaraniewski/openusage/internal/providers/providerbase"
-	"github.com/janekbaraniewski/openusage/internal/providers/shared"
-	"github.com/samber/lo"
 )
 
 const (
@@ -391,945 +388,6 @@ func (p *Provider) fetchLocalData(acct core.AccountConfig, snap *core.UsageSnaps
 	p.readSessions(copilotDir, snap, logData)
 }
 
-func (p *Provider) readConfig(copilotDir string, snap *core.UsageSnapshot) {
-	data, err := os.ReadFile(filepath.Join(copilotDir, "config.json"))
-	if err != nil {
-		return
-	}
-	var cfg copilotConfig
-	if json.Unmarshal(data, &cfg) != nil {
-		return
-	}
-	if cfg.Model != "" {
-		snap.Raw["preferred_model"] = cfg.Model
-	}
-	if cfg.ReasoningEffort != "" {
-		snap.Raw["reasoning_effort"] = cfg.ReasoningEffort
-	}
-	if cfg.Experimental {
-		snap.Raw["experimental"] = "enabled"
-	}
-}
-
-type logSummary struct {
-	DefaultModel  string
-	SessionTokens map[string]logTokenEntry // sessionID → last CompactionProcessor entry
-	SessionBurn   map[string]float64       // sessionID → cumulative positive token deltas
-}
-
-func (p *Provider) readLogs(copilotDir string, snap *core.UsageSnapshot) logSummary {
-	ls := logSummary{
-		SessionTokens: make(map[string]logTokenEntry),
-		SessionBurn:   make(map[string]float64),
-	}
-	sessionEntries := make(map[string][]logTokenEntry)
-	logDir := filepath.Join(copilotDir, "logs")
-	entries, err := os.ReadDir(logDir)
-	if err != nil {
-		return ls
-	}
-
-	var allTokenEntries []logTokenEntry
-
-	for _, entry := range entries {
-		if entry.IsDir() || !strings.HasSuffix(entry.Name(), ".log") {
-			continue
-		}
-		data, err := os.ReadFile(filepath.Join(logDir, entry.Name()))
-		if err != nil {
-			continue
-		}
-
-		var currentSessionID string
-		for _, line := range strings.Split(string(data), "\n") {
-			line = strings.TrimSpace(line)
-
-			if strings.Contains(line, "Workspace initialized:") {
-				if idx := strings.Index(line, "Workspace initialized:"); idx >= 0 {
-					rest := strings.TrimSpace(line[idx+len("Workspace initialized:"):])
-					if spIdx := strings.Index(rest, " "); spIdx > 0 {
-						currentSessionID = rest[:spIdx]
-					} else if rest != "" {
-						currentSessionID = rest
-					}
-				}
-			}
-
-			if strings.Contains(line, "Using default model:") {
-				if idx := strings.Index(line, "Using default model:"); idx >= 0 {
-					m := strings.TrimSpace(line[idx+len("Using default model:"):])
-					if m != "" {
-						ls.DefaultModel = m
-					}
-				}
-			}
-
-			if strings.Contains(line, "CompactionProcessor: Utilization") {
-				te := parseCompactionLine(line)
-				if te.Total > 0 {
-					allTokenEntries = append(allTokenEntries, te)
-					if currentSessionID != "" {
-						sessionEntries[currentSessionID] = append(sessionEntries[currentSessionID], te)
-					}
-				}
-			}
-		}
-	}
-
-	if ls.DefaultModel != "" {
-		snap.Raw["default_model"] = ls.DefaultModel
-	}
-
-	for sessionID, entries := range sessionEntries {
-		sortCompactionEntries(entries)
-		last := entries[len(entries)-1]
-		ls.SessionTokens[sessionID] = last
-
-		burn := 0.0
-		for idx, te := range entries {
-			if idx == 0 {
-				if te.Used > 0 {
-					burn += float64(te.Used)
-				}
-				continue
-			}
-			delta := te.Used - entries[idx-1].Used
-			if delta > 0 {
-				burn += float64(delta)
-			}
-		}
-		if burn > 0 {
-			ls.SessionBurn[sessionID] = burn
-		}
-	}
-
-	if last, ok := newestCompactionEntry(allTokenEntries); ok {
-		snap.Raw["context_window_tokens"] = fmt.Sprintf("%d/%d", last.Used, last.Total)
-		pct := float64(last.Used) / float64(last.Total) * 100
-		snap.Raw["context_window_pct"] = fmt.Sprintf("%.1f%%", pct)
-		used := float64(last.Used)
-		limit := float64(last.Total)
-		snap.Metrics["context_window"] = core.Metric{
-			Limit:     &limit,
-			Used:      &used,
-			Remaining: core.Float64Ptr(limit - used),
-			Unit:      "tokens",
-			Window:    "session",
-		}
-	}
-
-	return ls
-}
-
-type assistantMsgData struct {
-	Content      string          `json:"content"`
-	ReasoningTxt string          `json:"reasoningText"`
-	ToolRequests json.RawMessage `json:"toolRequests"`
-}
-
-type quotaSnapshotEntry struct {
-	EntitlementRequests int     `json:"entitlementRequests"`
-	UsedRequests        int     `json:"usedRequests"`
-	RemainingPercentage float64 `json:"remainingPercentage"`
-	ResetDate           string  `json:"resetDate"`
-}
-
-type assistantUsageData struct {
-	Model            string                        `json:"model"`
-	InputTokens      float64                       `json:"inputTokens"`
-	OutputTokens     float64                       `json:"outputTokens"`
-	CacheReadTokens  float64                       `json:"cacheReadTokens"`
-	CacheWriteTokens float64                       `json:"cacheWriteTokens"`
-	Cost             float64                       `json:"cost"`
-	Duration         int64                         `json:"duration"`
-	QuotaSnapshots   map[string]quotaSnapshotEntry `json:"quotaSnapshots"`
-}
-
-type sessionShutdownData struct {
-	ShutdownType         string                         `json:"shutdownType"`
-	TotalPremiumRequests int                            `json:"totalPremiumRequests"`
-	TotalAPIDurationMs   int64                          `json:"totalApiDurationMs"`
-	SessionStartTime     string                         `json:"sessionStartTime"`
-	CodeChanges          shutdownCodeChanges            `json:"codeChanges"`
-	ModelMetrics         map[string]shutdownModelMetric `json:"modelMetrics"`
-}
-
-type shutdownCodeChanges struct {
-	LinesAdded    int `json:"linesAdded"`
-	LinesRemoved  int `json:"linesRemoved"`
-	FilesModified int `json:"filesModified"`
-}
-
-type shutdownModelMetric struct {
-	Requests struct {
-		Count int     `json:"count"`
-		Cost  float64 `json:"cost"`
-	} `json:"requests"`
-	Usage struct {
-		InputTokens      float64 `json:"inputTokens"`
-		OutputTokens     float64 `json:"outputTokens"`
-		CacheReadTokens  float64 `json:"cacheReadTokens"`
-		CacheWriteTokens float64 `json:"cacheWriteTokens"`
-	} `json:"usage"`
-}
-
-func (p *Provider) readSessions(copilotDir string, snap *core.UsageSnapshot, logs logSummary) {
-	sessionDir := filepath.Join(copilotDir, "session-state")
-	entries, err := os.ReadDir(sessionDir)
-	if err != nil {
-		return
-	}
-
-	snap.Raw["total_sessions"] = fmt.Sprintf("%d", len(entries))
-
-	type sessionInfo struct {
-		id                      string
-		createdAt               time.Time
-		updatedAt               time.Time
-		cwd                     string
-		repo                    string
-		branch                  string
-		client                  string
-		summary                 string
-		messages                int
-		turns                   int
-		model                   string
-		responseChars           int
-		reasoningChars          int
-		toolCalls               int
-		tokenUsed               int
-		tokenTotal              int
-		tokenBurn               float64
-		usageCost               float64
-		premiumRequests         int
-		shutdownPremiumRequests int
-		linesAdded              int
-		linesRemoved            int
-		filesModified           int
-	}
-
-	var sessions []sessionInfo
-	dailyMessages := make(map[string]float64)
-	dailySessions := make(map[string]float64)
-	dailyToolCalls := make(map[string]float64)
-	dailyTokens := make(map[string]float64)
-	modelMessages := make(map[string]int)
-	modelTurns := make(map[string]int)
-	modelSessions := make(map[string]int)
-	modelResponseChars := make(map[string]int)
-	modelReasoningChars := make(map[string]int)
-	modelToolCalls := make(map[string]int)
-	dailyModelMessages := make(map[string]map[string]float64)
-	dailyModelTokens := make(map[string]map[string]float64)
-	modelInputTokens := make(map[string]float64)
-	usageInputTokens := make(map[string]float64)
-	usageOutputTokens := make(map[string]float64)
-	usageCacheReadTokens := make(map[string]float64)
-	usageCacheWriteTokens := make(map[string]float64)
-	usageCost := make(map[string]float64)
-	usageRequests := make(map[string]int)
-	usageDuration := make(map[string]int64)
-	dailyCost := make(map[string]float64)
-	var latestQuotaSnapshots map[string]quotaSnapshotEntry
-	var shutdownPremiumRequests int
-	var shutdownLinesAdded, shutdownLinesRemoved, shutdownFilesModified int
-	shutdownModelCost := make(map[string]float64)
-	shutdownModelRequests := make(map[string]int)
-	shutdownModelInputTokens := make(map[string]float64)
-	shutdownModelOutputTokens := make(map[string]float64)
-	toolUsageCounts := make(map[string]int)
-	languageUsageCounts := make(map[string]int)
-	changedFiles := make(map[string]bool)
-	commitCommands := make(map[string]bool)
-	clientLabels := make(map[string]string)
-	clientTokens := make(map[string]float64)
-	clientSessions := make(map[string]int)
-	clientMessages := make(map[string]int)
-	dailyClientTokens := make(map[string]map[string]float64)
-	var inferredLinesAdded, inferredLinesRemoved int
-	var inferredCommitCount int
-
-	for _, entry := range entries {
-		if !entry.IsDir() {
-			continue
-		}
-		si := sessionInfo{id: entry.Name()}
-		sessPath := filepath.Join(sessionDir, entry.Name())
-
-		if wsData, err := os.ReadFile(filepath.Join(sessPath, "workspace.yaml")); err == nil {
-			ws := parseSimpleYAML(string(wsData))
-			si.cwd = ws["cwd"]
-			si.repo = ws["repository"]
-			si.branch = ws["branch"]
-			si.summary = ws["summary"]
-			si.createdAt = flexParseTime(ws["created_at"])
-			si.updatedAt = flexParseTime(ws["updated_at"])
-		}
-
-		if te, ok := logs.SessionTokens[si.id]; ok {
-			si.tokenUsed = te.Used
-			si.tokenTotal = te.Total
-			if !te.Timestamp.IsZero() {
-				if si.createdAt.IsZero() {
-					si.createdAt = te.Timestamp
-				}
-				if si.updatedAt.IsZero() || te.Timestamp.After(si.updatedAt) {
-					si.updatedAt = te.Timestamp
-				}
-			}
-		}
-		if burn, ok := logs.SessionBurn[si.id]; ok {
-			si.tokenBurn = burn
-		}
-
-		if evtData, err := os.ReadFile(filepath.Join(sessPath, "events.jsonl")); err == nil {
-			currentModel := logs.DefaultModel
-			var firstEventAt, lastEventAt time.Time
-			lines := strings.Split(string(evtData), "\n")
-			for _, line := range lines {
-				line = strings.TrimSpace(line)
-				if line == "" {
-					continue
-				}
-				var evt sessionEvent
-				if json.Unmarshal([]byte(line), &evt) != nil {
-					continue
-				}
-				evtTime := flexParseTime(evt.Timestamp)
-				if !evtTime.IsZero() {
-					if firstEventAt.IsZero() || evtTime.Before(firstEventAt) {
-						firstEventAt = evtTime
-					}
-					if lastEventAt.IsZero() || evtTime.After(lastEventAt) {
-						lastEventAt = evtTime
-					}
-				}
-
-				switch evt.Type {
-				case "session.start":
-					var start sessionStartData
-					if json.Unmarshal(evt.Data, &start) == nil {
-						if si.cwd == "" {
-							si.cwd = start.Context.CWD
-						}
-						if si.repo == "" {
-							si.repo = start.Context.Repository
-						}
-						if si.branch == "" {
-							si.branch = start.Context.Branch
-						}
-						if si.createdAt.IsZero() {
-							si.createdAt = flexParseTime(start.StartTime)
-						}
-						if currentModel == "" && start.SelectedModel != "" {
-							currentModel = start.SelectedModel
-						}
-					}
-
-				case "session.model_change":
-					var mc modelChangeData
-					if json.Unmarshal(evt.Data, &mc) == nil && mc.NewModel != "" {
-						currentModel = mc.NewModel
-					}
-
-				case "session.info":
-					var info sessionInfoData
-					if json.Unmarshal(evt.Data, &info) == nil && info.InfoType == "model" {
-						if m := extractModelFromInfoMsg(info.Message); m != "" {
-							currentModel = m
-						}
-					}
-
-				case "user.message":
-					si.messages++
-					day := parseDayFromTimestamp(evt.Timestamp)
-					if day != "" {
-						dailyMessages[day]++
-					}
-					if currentModel != "" {
-						modelMessages[currentModel]++
-						if day != "" {
-							if dailyModelMessages[currentModel] == nil {
-								dailyModelMessages[currentModel] = make(map[string]float64)
-							}
-							dailyModelMessages[currentModel][day]++
-						}
-					}
-
-				case "assistant.turn_start":
-					si.turns++
-					if currentModel != "" {
-						modelTurns[currentModel]++
-					}
-
-				case "assistant.message":
-					var msg assistantMsgData
-					if json.Unmarshal(evt.Data, &msg) == nil {
-						si.responseChars += len(msg.Content)
-						si.reasoningChars += len(msg.ReasoningTxt)
-						if currentModel != "" {
-							modelResponseChars[currentModel] += len(msg.Content)
-							modelReasoningChars[currentModel] += len(msg.ReasoningTxt)
-						}
-						var tools []json.RawMessage
-						if json.Unmarshal(msg.ToolRequests, &tools) == nil && len(tools) > 0 {
-							si.toolCalls += len(tools)
-							if currentModel != "" {
-								modelToolCalls[currentModel] += len(tools)
-							}
-							for _, toolReq := range tools {
-								toolName := extractCopilotToolName(toolReq)
-								if toolName == "" {
-									toolName = "unknown"
-								}
-								toolUsageCounts[toolName]++
-								toolLower := strings.ToLower(strings.TrimSpace(toolName))
-								paths := extractCopilotToolPaths(toolReq)
-								for _, path := range paths {
-									if lang := inferCopilotLanguageFromPath(path); lang != "" {
-										languageUsageCounts[lang]++
-									}
-									if isCopilotMutatingTool(toolLower) {
-										changedFiles[path] = true
-									}
-								}
-								if isCopilotMutatingTool(toolLower) {
-									added, removed := estimateCopilotToolLineDelta(toolReq)
-									inferredLinesAdded += added
-									inferredLinesRemoved += removed
-								}
-								cmd := extractCopilotToolCommand(toolReq)
-								if cmd != "" {
-									if strings.Contains(strings.ToLower(cmd), "git commit") && !commitCommands[cmd] {
-										commitCommands[cmd] = true
-										inferredCommitCount++
-									}
-								} else if strings.Contains(toolLower, "commit") {
-									inferredCommitCount++
-								}
-							}
-							day := parseDayFromTimestamp(evt.Timestamp)
-							if day != "" {
-								dailyToolCalls[day] += float64(len(tools))
-							}
-						}
-					}
-
-				case "assistant.usage":
-					var usage assistantUsageData
-					if json.Unmarshal(evt.Data, &usage) == nil && usage.Model != "" {
-						usageInputTokens[usage.Model] += usage.InputTokens
-						usageOutputTokens[usage.Model] += usage.OutputTokens
-						usageCacheReadTokens[usage.Model] += usage.CacheReadTokens
-						usageCacheWriteTokens[usage.Model] += usage.CacheWriteTokens
-						usageCost[usage.Model] += usage.Cost
-						usageRequests[usage.Model]++
-						usageDuration[usage.Model] += usage.Duration
-
-						si.usageCost += usage.Cost
-						si.premiumRequests++
-
-						day := parseDayFromTimestamp(evt.Timestamp)
-						if day != "" {
-							dailyCost[day] += usage.Cost
-						}
-
-						if len(usage.QuotaSnapshots) > 0 {
-							latestQuotaSnapshots = usage.QuotaSnapshots
-						}
-					}
-
-				case "session.shutdown":
-					var shutdown sessionShutdownData
-					if json.Unmarshal(evt.Data, &shutdown) == nil {
-						shutdownPremiumRequests += shutdown.TotalPremiumRequests
-						si.shutdownPremiumRequests += shutdown.TotalPremiumRequests
-
-						si.linesAdded += shutdown.CodeChanges.LinesAdded
-						si.linesRemoved += shutdown.CodeChanges.LinesRemoved
-						si.filesModified += shutdown.CodeChanges.FilesModified
-						shutdownLinesAdded += shutdown.CodeChanges.LinesAdded
-						shutdownLinesRemoved += shutdown.CodeChanges.LinesRemoved
-						shutdownFilesModified += shutdown.CodeChanges.FilesModified
-
-						for model, metrics := range shutdown.ModelMetrics {
-							shutdownModelCost[model] += metrics.Requests.Cost
-							shutdownModelRequests[model] += metrics.Requests.Count
-							shutdownModelInputTokens[model] += metrics.Usage.InputTokens
-							shutdownModelOutputTokens[model] += metrics.Usage.OutputTokens
-						}
-					}
-				}
-			}
-			if !firstEventAt.IsZero() && si.createdAt.IsZero() {
-				si.createdAt = firstEventAt
-			}
-			if !lastEventAt.IsZero() && (si.updatedAt.IsZero() || lastEventAt.After(si.updatedAt)) {
-				si.updatedAt = lastEventAt
-			}
-			si.model = currentModel
-		}
-
-		day := dayForSession(si.createdAt, si.updatedAt)
-		if si.model != "" {
-			modelSessions[si.model]++
-		}
-		if day != "" {
-			dailySessions[day]++
-		}
-
-		clientLabel := normalizeCopilotClient(si.repo, si.cwd)
-		clientKey := sanitizeMetricName(clientLabel)
-		if clientKey == "" {
-			clientKey = "cli"
-		}
-		si.client = clientLabel
-		if _, ok := clientLabels[clientKey]; !ok {
-			clientLabels[clientKey] = clientLabel
-		}
-		clientSessions[clientKey]++
-		clientMessages[clientKey] += si.messages
-
-		sessionTokens := float64(si.tokenUsed)
-		if si.tokenBurn > 0 {
-			sessionTokens = si.tokenBurn
-		}
-		if sessionTokens > 0 {
-			clientTokens[clientKey] += sessionTokens
-			if day != "" {
-				dailyTokens[day] += sessionTokens
-				if dailyClientTokens[clientKey] == nil {
-					dailyClientTokens[clientKey] = make(map[string]float64)
-				}
-				dailyClientTokens[clientKey][day] += sessionTokens
-			}
-			if si.model != "" {
-				modelInputTokens[si.model] += sessionTokens
-				if day != "" {
-					if dailyModelTokens[si.model] == nil {
-						dailyModelTokens[si.model] = make(map[string]float64)
-					}
-					dailyModelTokens[si.model][day] += sessionTokens
-				}
-			}
-		}
-		sessions = append(sessions, si)
-	}
-
-	storeSeries(snap, "messages", dailyMessages)
-	storeSeries(snap, "sessions", dailySessions)
-	storeSeries(snap, "tool_calls", dailyToolCalls)
-	storeSeries(snap, "tokens_total", dailyTokens)
-	storeSeries(snap, "cli_messages", dailyMessages)
-	storeSeries(snap, "cli_sessions", dailySessions)
-	storeSeries(snap, "cli_tool_calls", dailyToolCalls)
-	if len(dailyCost) > 0 {
-		storeSeries(snap, "cost", dailyCost)
-	}
-	for model, dayCounts := range dailyModelMessages {
-		safe := sanitizeMetricName(model)
-		storeSeries(snap, "cli_messages_"+safe, dayCounts)
-	}
-	for model, dayCounts := range dailyModelTokens {
-		safe := sanitizeMetricName(model)
-		storeSeries(snap, "tokens_"+safe, dayCounts)
-		storeSeries(snap, "cli_tokens_"+safe, dayCounts)
-	}
-
-	setRawStr(snap, "model_usage", formatModelMap(modelMessages, "msgs"))
-	setRawStr(snap, "model_turns", formatModelMap(modelTurns, "turns"))
-	setRawStr(snap, "model_sessions", formatModelMapPlain(modelSessions))
-	setRawStr(snap, "model_response_chars", formatModelMap(modelResponseChars, "chars"))
-	setRawStr(snap, "model_reasoning_chars", formatModelMap(modelReasoningChars, "chars"))
-	setRawStr(snap, "model_tool_calls", formatModelMap(modelToolCalls, "calls"))
-
-	sort.Slice(sessions, func(i, j int) bool {
-		ti := sessions[i].updatedAt
-		if ti.IsZero() {
-			ti = sessions[i].createdAt
-		}
-		tj := sessions[j].updatedAt
-		if tj.IsZero() {
-			tj = sessions[j].createdAt
-		}
-		return ti.After(tj)
-	})
-
-	var totalMessages, totalTurns, totalResponse, totalReasoning, totalTools int
-	totalTokens := 0.0
-	for _, s := range sessions {
-		totalMessages += s.messages
-		totalTurns += s.turns
-		totalResponse += s.responseChars
-		totalReasoning += s.reasoningChars
-		totalTools += s.toolCalls
-		tokens := float64(s.tokenUsed)
-		if s.tokenBurn > 0 {
-			tokens = s.tokenBurn
-		}
-		totalTokens += tokens
-	}
-	setRawInt(snap, "total_cli_messages", totalMessages)
-	setRawInt(snap, "total_cli_turns", totalTurns)
-	setRawInt(snap, "total_response_chars", totalResponse)
-	setRawInt(snap, "total_reasoning_chars", totalReasoning)
-	setRawInt(snap, "total_tool_calls", totalTools)
-
-	setUsedMetric(snap, "total_messages", float64(totalMessages), "messages", copilotAllTimeWindow)
-	setUsedMetric(snap, "total_sessions", float64(len(sessions)), "sessions", copilotAllTimeWindow)
-	setUsedMetric(snap, "total_turns", float64(totalTurns), "turns", copilotAllTimeWindow)
-	setUsedMetric(snap, "total_tool_calls", float64(totalTools), "calls", copilotAllTimeWindow)
-	setUsedMetric(snap, "tool_calls_total", float64(totalTools), "calls", copilotAllTimeWindow)
-	if totalTools > 0 {
-		setUsedMetric(snap, "tool_completed", float64(totalTools), "calls", copilotAllTimeWindow)
-		setUsedMetric(snap, "tool_success_rate", 100.0, "%", copilotAllTimeWindow)
-	}
-	setUsedMetric(snap, "total_response_chars", float64(totalResponse), "chars", copilotAllTimeWindow)
-	setUsedMetric(snap, "total_reasoning_chars", float64(totalReasoning), "chars", copilotAllTimeWindow)
-	setUsedMetric(snap, "total_conversations", float64(len(sessions)), "sessions", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_messages", float64(totalMessages), "messages", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_turns", float64(totalTurns), "turns", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_sessions", float64(len(sessions)), "sessions", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_tool_calls", float64(totalTools), "calls", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_response_chars", float64(totalResponse), "chars", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_reasoning_chars", float64(totalReasoning), "chars", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_input_tokens", totalTokens, "tokens", copilotAllTimeWindow)
-	setUsedMetric(snap, "cli_total_tokens", totalTokens, "tokens", copilotAllTimeWindow)
-
-	// Emit new metrics from assistant.usage and session.shutdown events.
-	var totalUsageOutputTokens, totalUsageCacheRead, totalUsageCacheWrite, totalUsageCost float64
-	var totalUsageRequests int
-	for _, v := range usageOutputTokens {
-		totalUsageOutputTokens += v
-	}
-	for _, v := range usageCacheReadTokens {
-		totalUsageCacheRead += v
-	}
-	for _, v := range usageCacheWriteTokens {
-		totalUsageCacheWrite += v
-	}
-	for _, v := range usageCost {
-		totalUsageCost += v
-	}
-	for _, v := range usageRequests {
-		totalUsageRequests += v
-	}
-	if totalUsageOutputTokens > 0 {
-		setUsedMetric(snap, "cli_output_tokens", totalUsageOutputTokens, "tokens", copilotAllTimeWindow)
-	}
-	if totalUsageCacheRead > 0 {
-		setUsedMetric(snap, "cli_cache_read_tokens", totalUsageCacheRead, "tokens", copilotAllTimeWindow)
-	}
-	if totalUsageCacheWrite > 0 {
-		setUsedMetric(snap, "cli_cache_write_tokens", totalUsageCacheWrite, "tokens", copilotAllTimeWindow)
-	}
-	if totalUsageCost > 0 {
-		setUsedMetric(snap, "cli_cost", totalUsageCost, "USD", copilotAllTimeWindow)
-	}
-	if totalUsageRequests > 0 {
-		setUsedMetric(snap, "cli_premium_requests", float64(totalUsageRequests), "requests", copilotAllTimeWindow)
-	} else if shutdownPremiumRequests > 0 {
-		setUsedMetric(snap, "cli_premium_requests", float64(shutdownPremiumRequests), "requests", copilotAllTimeWindow)
-	}
-	if shutdownLinesAdded > 0 || shutdownLinesRemoved > 0 {
-		setUsedMetric(snap, "cli_lines_added", float64(shutdownLinesAdded), "lines", copilotAllTimeWindow)
-		setUsedMetric(snap, "cli_lines_removed", float64(shutdownLinesRemoved), "lines", copilotAllTimeWindow)
-	}
-	if shutdownFilesModified > 0 {
-		setUsedMetric(snap, "cli_files_modified", float64(shutdownFilesModified), "files", copilotAllTimeWindow)
-	}
-	if totalUsageRequests > 0 {
-		var totalDuration int64
-		for _, d := range usageDuration {
-			totalDuration += d
-		}
-		avgMs := float64(totalDuration) / float64(totalUsageRequests)
-		setUsedMetric(snap, "cli_avg_latency_ms", avgMs, "ms", copilotAllTimeWindow)
-	}
-
-	// Apply latestQuotaSnapshots as fallback for premium_interactions_quota.
-	if qs, ok := latestQuotaSnapshots["premium_interactions"]; ok {
-		if _, exists := snap.Metrics["premium_interactions_quota"]; !exists {
-			entitlement := float64(qs.EntitlementRequests)
-			used := float64(qs.UsedRequests)
-			remaining := entitlement - used
-			if remaining < 0 {
-				remaining = 0
-			}
-			snap.Metrics["premium_interactions_quota"] = core.Metric{
-				Limit:     &entitlement,
-				Used:      core.Float64Ptr(used),
-				Remaining: core.Float64Ptr(remaining),
-				Unit:      "requests",
-				Window:    "billing-cycle",
-			}
-		}
-	}
-
-	if _, v := latestSeriesValue(dailyCost); v > 0 {
-		setUsedMetric(snap, "cost_today", v, "USD", "today")
-	}
-	setUsedMetric(snap, "7d_cost", sumLastNDays(dailyCost, 7), "USD", "7d")
-
-	if _, v := latestSeriesValue(dailyMessages); v > 0 {
-		setUsedMetric(snap, "messages_today", v, "messages", "today")
-	}
-	if _, v := latestSeriesValue(dailySessions); v > 0 {
-		setUsedMetric(snap, "sessions_today", v, "sessions", "today")
-	}
-	if _, v := latestSeriesValue(dailyToolCalls); v > 0 {
-		setUsedMetric(snap, "tool_calls_today", v, "calls", "today")
-	}
-	if _, v := latestSeriesValue(dailyTokens); v > 0 {
-		setUsedMetric(snap, "tokens_today", v, "tokens", "today")
-	}
-	setUsedMetric(snap, "7d_messages", sumLastNDays(dailyMessages, 7), "messages", "7d")
-	setUsedMetric(snap, "7d_sessions", sumLastNDays(dailySessions, 7), "sessions", "7d")
-	setUsedMetric(snap, "7d_tool_calls", sumLastNDays(dailyToolCalls, 7), "calls", "7d")
-	setUsedMetric(snap, "7d_tokens", sumLastNDays(dailyTokens, 7), "tokens", "7d")
-	setUsedMetric(snap, "total_prompts", float64(totalMessages), "prompts", copilotAllTimeWindow)
-
-	// Merge usage event models into the topModels set so they appear even if
-	// they have no log-compaction data.
-	allModelTokens := make(map[string]float64, len(modelInputTokens))
-	for k, v := range modelInputTokens {
-		allModelTokens[k] = v
-	}
-	for k, v := range usageInputTokens {
-		if allModelTokens[k] < v {
-			allModelTokens[k] = v
-		}
-	}
-	allModelMessages := make(map[string]int, len(modelMessages))
-	for k, v := range modelMessages {
-		allModelMessages[k] = v
-	}
-	for k, v := range usageRequests {
-		if allModelMessages[k] < v {
-			allModelMessages[k] = v
-		}
-	}
-	topModels := topModelNames(allModelTokens, allModelMessages, maxCopilotModels)
-	for _, model := range topModels {
-		prefix := "model_" + sanitizeMetricName(model)
-		rec := core.ModelUsageRecord{
-			RawModelID: model,
-			RawSource:  "json",
-			Window:     copilotAllTimeWindow,
-		}
-
-		// Prefer usage event data (accurate) over log-compaction data (approximate).
-		inputTok := modelInputTokens[model]
-		if v := usageInputTokens[model]; v > 0 {
-			inputTok = v
-		}
-		outputTok := usageOutputTokens[model]
-		cacheTok := usageCacheReadTokens[model] + usageCacheWriteTokens[model]
-
-		setUsedMetric(snap, prefix+"_input_tokens", inputTok, "tokens", copilotAllTimeWindow)
-		if inputTok > 0 {
-			rec.InputTokens = core.Float64Ptr(inputTok)
-		}
-		if outputTok > 0 {
-			setUsedMetric(snap, prefix+"_output_tokens", outputTok, "tokens", copilotAllTimeWindow)
-			rec.OutputTokens = core.Float64Ptr(outputTok)
-		}
-		if cacheTok > 0 {
-			rec.CachedTokens = core.Float64Ptr(cacheTok)
-		}
-		totalTok := inputTok + outputTok
-		if totalTok > 0 {
-			rec.TotalTokens = core.Float64Ptr(totalTok)
-		}
-
-		// Cost from usage events; fall back to shutdown model metrics.
-		modelCost := usageCost[model]
-		if modelCost == 0 {
-			modelCost = shutdownModelCost[model]
-		}
-		if modelCost > 0 {
-			rec.CostUSD = core.Float64Ptr(modelCost)
-			setUsedMetric(snap, prefix+"_cost", modelCost, "USD", copilotAllTimeWindow)
-		}
-
-		// Requests from usage events.
-		if reqs := usageRequests[model]; reqs > 0 {
-			rec.Requests = core.Float64Ptr(float64(reqs))
-		}
-
-		setUsedMetric(snap, prefix+"_messages", float64(modelMessages[model]), "messages", copilotAllTimeWindow)
-		setUsedMetric(snap, prefix+"_turns", float64(modelTurns[model]), "turns", copilotAllTimeWindow)
-		setUsedMetric(snap, prefix+"_sessions", float64(modelSessions[model]), "sessions", copilotAllTimeWindow)
-		setUsedMetric(snap, prefix+"_tool_calls", float64(modelToolCalls[model]), "calls", copilotAllTimeWindow)
-		setUsedMetric(snap, prefix+"_response_chars", float64(modelResponseChars[model]), "chars", copilotAllTimeWindow)
-		setUsedMetric(snap, prefix+"_reasoning_chars", float64(modelReasoningChars[model]), "chars", copilotAllTimeWindow)
-		snap.AppendModelUsage(rec)
-	}
-
-	topClients := topCopilotClientNames(clientTokens, clientSessions, clientMessages, maxCopilotClients)
-	for _, client := range topClients {
-		clientPrefix := "client_" + client
-		setUsedMetric(snap, clientPrefix+"_total_tokens", clientTokens[client], "tokens", copilotAllTimeWindow)
-		setUsedMetric(snap, clientPrefix+"_input_tokens", clientTokens[client], "tokens", copilotAllTimeWindow)
-		setUsedMetric(snap, clientPrefix+"_sessions", float64(clientSessions[client]), "sessions", copilotAllTimeWindow)
-		if byDay := dailyClientTokens[client]; len(byDay) > 0 {
-			storeSeries(snap, "tokens_client_"+client, byDay)
-		}
-	}
-	setRawStr(snap, "client_usage", formatCopilotClientUsage(topClients, clientLabels, clientTokens, clientSessions))
-	setRawStr(snap, "tool_usage", formatModelMap(toolUsageCounts, "calls"))
-	setRawStr(snap, "language_usage", formatModelMap(languageUsageCounts, "req"))
-	for toolName, count := range toolUsageCounts {
-		if count <= 0 {
-			continue
-		}
-		setUsedMetric(snap, "tool_"+sanitizeMetricName(toolName), float64(count), "calls", copilotAllTimeWindow)
-	}
-	for lang, count := range languageUsageCounts {
-		if count <= 0 {
-			continue
-		}
-		setUsedMetric(snap, "lang_"+sanitizeMetricName(lang), float64(count), "requests", copilotAllTimeWindow)
-	}
-
-	linesAdded := shutdownLinesAdded
-	if inferredLinesAdded > linesAdded {
-		linesAdded = inferredLinesAdded
-	}
-	linesRemoved := shutdownLinesRemoved
-	if inferredLinesRemoved > linesRemoved {
-		linesRemoved = inferredLinesRemoved
-	}
-	filesChanged := shutdownFilesModified
-	if len(changedFiles) > filesChanged {
-		filesChanged = len(changedFiles)
-	}
-	if linesAdded > 0 {
-		setUsedMetric(snap, "composer_lines_added", float64(linesAdded), "lines", copilotAllTimeWindow)
-	}
-	if linesRemoved > 0 {
-		setUsedMetric(snap, "composer_lines_removed", float64(linesRemoved), "lines", copilotAllTimeWindow)
-	}
-	if filesChanged > 0 {
-		setUsedMetric(snap, "composer_files_changed", float64(filesChanged), "files", copilotAllTimeWindow)
-	}
-	if inferredCommitCount > 0 {
-		setUsedMetric(snap, "scored_commits", float64(inferredCommitCount), "commits", copilotAllTimeWindow)
-	}
-	if linesAdded > 0 || linesRemoved > 0 {
-		hundred := 100.0
-		zero := 0.0
-		snap.Metrics["ai_code_percentage"] = core.Metric{
-			Used:      &hundred,
-			Remaining: &zero,
-			Limit:     &hundred,
-			Unit:      "%",
-			Window:    copilotAllTimeWindow,
-		}
-	}
-
-	if len(sessions) > 0 {
-		r := sessions[0]
-		if r.client != "" {
-			snap.Raw["last_session_client"] = r.client
-		}
-		snap.Raw["last_session_repo"] = r.repo
-		snap.Raw["last_session_branch"] = r.branch
-		if r.summary != "" {
-			snap.Raw["last_session_summary"] = r.summary
-		}
-		if !r.updatedAt.IsZero() {
-			snap.Raw["last_session_time"] = r.updatedAt.Format(time.RFC3339)
-		}
-		if r.model != "" {
-			snap.Raw["last_session_model"] = r.model
-		}
-		sessionTokens := float64(r.tokenUsed)
-		if r.tokenBurn > 0 {
-			sessionTokens = r.tokenBurn
-		}
-		if sessionTokens > 0 {
-			snap.Raw["last_session_tokens"] = fmt.Sprintf("%.0f/%d", sessionTokens, r.tokenTotal)
-			setUsedMetric(snap, "session_input_tokens", sessionTokens, "tokens", "session")
-			setUsedMetric(snap, "session_total_tokens", sessionTokens, "tokens", "session")
-			if r.tokenTotal > 0 {
-				limit := float64(r.tokenTotal)
-				snap.Metrics["context_window"] = core.Metric{
-					Limit:     &limit,
-					Used:      core.Float64Ptr(sessionTokens),
-					Remaining: core.Float64Ptr(max(limit-sessionTokens, 0)),
-					Unit:      "tokens",
-					Window:    "session",
-				}
-			}
-		}
-	}
-}
-
-func parseCompactionLine(line string) logTokenEntry {
-	var entry logTokenEntry
-
-	if len(line) >= 24 {
-		if t, err := time.Parse("2006-01-02T15:04:05.000Z", line[:24]); err == nil {
-			entry.Timestamp = t
-		}
-	}
-
-	parenStart := strings.Index(line, "(")
-	parenEnd := strings.Index(line, " tokens)")
-	if parenStart >= 0 && parenEnd > parenStart {
-		inner := line[parenStart+1 : parenEnd]
-		parts := strings.Split(inner, "/")
-		if len(parts) == 2 {
-			fmt.Sscanf(parts[0], "%d", &entry.Used)
-			fmt.Sscanf(parts[1], "%d", &entry.Total)
-		}
-	}
-
-	return entry
-}
-
-func sortCompactionEntries(entries []logTokenEntry) {
-	sort.SliceStable(entries, func(i, j int) bool {
-		ti := entries[i].Timestamp
-		tj := entries[j].Timestamp
-		switch {
-		case ti.IsZero() && tj.IsZero():
-			return entries[i].Used < entries[j].Used
-		case ti.IsZero():
-			return false
-		case tj.IsZero():
-			return true
-		default:
-			return ti.Before(tj)
-		}
-	})
-}
-
-func newestCompactionEntry(entries []logTokenEntry) (logTokenEntry, bool) {
-	if len(entries) == 0 {
-		return logTokenEntry{}, false
-	}
-	best := entries[0]
-	for _, te := range entries[1:] {
-		if best.Timestamp.IsZero() && !te.Timestamp.IsZero() {
-			best = te
-			continue
-		}
-		if !best.Timestamp.IsZero() && te.Timestamp.IsZero() {
-			continue
-		}
-		if !te.Timestamp.IsZero() && te.Timestamp.After(best.Timestamp) {
-			best = te
-			continue
-		}
-		if best.Timestamp.Equal(te.Timestamp) && te.Used > best.Used {
-			best = te
-		}
-	}
-	return best, true
-}
-
 func (p *Provider) resolveStatus(snap *core.UsageSnapshot, authOutput string) {
 	lower := strings.ToLower(authOutput)
 	if strings.Contains(lower, "rate limit") || strings.Contains(lower, "rate_limit") {
@@ -1397,563 +455,6 @@ func skuLabel(sku string) string {
 	}
 }
 
-func parseSimpleYAML(content string) map[string]string {
-	result := make(map[string]string)
-	for _, line := range strings.Split(content, "\n") {
-		line = strings.TrimSpace(line)
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-		idx := strings.Index(line, ":")
-		if idx < 0 {
-			continue
-		}
-		key := strings.TrimSpace(line[:idx])
-		val := strings.TrimSpace(line[idx+1:])
-		result[key] = val
-	}
-	return result
-}
-
-func storeSeries(snap *core.UsageSnapshot, key string, m map[string]float64) {
-	if len(m) > 0 {
-		snap.DailySeries[key] = core.SortedTimePoints(m)
-	}
-}
-
-func setUsedMetric(snap *core.UsageSnapshot, key string, value float64, unit, window string) {
-	if value <= 0 {
-		return
-	}
-	v := value
-	snap.Metrics[key] = core.Metric{
-		Used:   &v,
-		Unit:   unit,
-		Window: window,
-	}
-}
-
-func dayForSession(createdAt, updatedAt time.Time) string {
-	if !updatedAt.IsZero() {
-		return updatedAt.Format("2006-01-02")
-	}
-	if !createdAt.IsZero() {
-		return createdAt.Format("2006-01-02")
-	}
-	return ""
-}
-
-func latestSeriesValue(m map[string]float64) (string, float64) {
-	if len(m) == 0 {
-		return "", 0
-	}
-	dates := lo.Keys(m)
-	sort.Strings(dates)
-	last := dates[len(dates)-1]
-	return last, m[last]
-}
-
-func sumLastNDays(m map[string]float64, days int) float64 {
-	if len(m) == 0 || days <= 0 {
-		return 0
-	}
-	date, _ := latestSeriesValue(m)
-	if date == "" {
-		return 0
-	}
-	end, err := time.Parse("2006-01-02", date)
-	if err != nil {
-		return 0
-	}
-	start := end.AddDate(0, 0, -(days - 1))
-	sum := 0.0
-	for d, v := range m {
-		t, err := time.Parse("2006-01-02", d)
-		if err != nil {
-			continue
-		}
-		if !t.Before(start) && !t.After(end) {
-			sum += v
-		}
-	}
-	return sum
-}
-
-func topModelNames(tokenMap map[string]float64, messageMap map[string]int, limit int) []string {
-	type row struct {
-		model    string
-		tokens   float64
-		messages int
-	}
-
-	seen := make(map[string]bool)
-	var rows []row
-	for model, tokens := range tokenMap {
-		seen[model] = true
-		rows = append(rows, row{model: model, tokens: tokens, messages: messageMap[model]})
-	}
-	for model, messages := range messageMap {
-		if seen[model] {
-			continue
-		}
-		rows = append(rows, row{model: model, messages: messages})
-	}
-
-	sort.Slice(rows, func(i, j int) bool {
-		if rows[i].tokens == rows[j].tokens {
-			if rows[i].messages == rows[j].messages {
-				return rows[i].model < rows[j].model
-			}
-			return rows[i].messages > rows[j].messages
-		}
-		return rows[i].tokens > rows[j].tokens
-	})
-
-	if limit > 0 && len(rows) > limit {
-		rows = rows[:limit]
-	}
-	return lo.Map(rows, func(r row, _ int) string { return r.model })
-}
-
-func topCopilotClientNames(tokenMap map[string]float64, sessionMap, messageMap map[string]int, limit int) []string {
-	type row struct {
-		client   string
-		tokens   float64
-		sessions int
-		messages int
-	}
-
-	seen := make(map[string]bool)
-	var rows []row
-	for client, tokens := range tokenMap {
-		seen[client] = true
-		rows = append(rows, row{
-			client:   client,
-			tokens:   tokens,
-			sessions: sessionMap[client],
-			messages: messageMap[client],
-		})
-	}
-	for client, sessions := range sessionMap {
-		if seen[client] {
-			continue
-		}
-		seen[client] = true
-		rows = append(rows, row{
-			client:   client,
-			sessions: sessions,
-			messages: messageMap[client],
-		})
-	}
-	for client, messages := range messageMap {
-		if seen[client] {
-			continue
-		}
-		rows = append(rows, row{
-			client:   client,
-			messages: messages,
-		})
-	}
-
-	sort.Slice(rows, func(i, j int) bool {
-		if rows[i].tokens == rows[j].tokens {
-			if rows[i].sessions == rows[j].sessions {
-				if rows[i].messages == rows[j].messages {
-					return rows[i].client < rows[j].client
-				}
-				return rows[i].messages > rows[j].messages
-			}
-			return rows[i].sessions > rows[j].sessions
-		}
-		return rows[i].tokens > rows[j].tokens
-	})
-
-	if limit > 0 && len(rows) > limit {
-		rows = rows[:limit]
-	}
-	return lo.Map(rows, func(r row, _ int) string { return r.client })
-}
-
-func normalizeCopilotClient(repo, cwd string) string {
-	repo = strings.TrimSpace(repo)
-	if repo != "" && repo != "." {
-		return repo
-	}
-
-	cwd = strings.TrimSpace(cwd)
-	if cwd != "" {
-		base := strings.TrimSpace(filepath.Base(cwd))
-		if base != "" && base != "." && base != string(filepath.Separator) {
-			return base
-		}
-	}
-
-	return "cli"
-}
-
-func formatCopilotClientUsage(clients []string, labels map[string]string, tokens map[string]float64, sessions map[string]int) string {
-	if len(clients) == 0 {
-		return ""
-	}
-
-	parts := make([]string, 0, len(clients))
-	for _, client := range clients {
-		label := labels[client]
-		if label == "" {
-			label = client
-		}
-
-		value := tokens[client]
-		sessionCount := sessions[client]
-
-		item := fmt.Sprintf("%s %s tok", label, formatCopilotTokenCount(value))
-		if sessionCount > 0 {
-			item += fmt.Sprintf(" · %d sess", sessionCount)
-		}
-		parts = append(parts, item)
-	}
-	return strings.Join(parts, ", ")
-}
-
-func formatCopilotTokenCount(value float64) string { return shared.FormatTokenCountF(value) }
-
-func parseDayFromTimestamp(ts string) string {
-	t := flexParseTime(ts)
-	if t.IsZero() {
-		return ""
-	}
-	return t.Format("2006-01-02")
-}
-
-func flexParseTime(s string) time.Time {
-	return shared.FlexParseTime(s)
-}
-
-func parseCopilotTime(s string) time.Time {
-	return shared.FlexParseTime(s)
-}
-
-func extractModelFromInfoMsg(msg string) string {
-	idx := strings.Index(msg, ": ")
-	if idx < 0 {
-		return ""
-	}
-	m := strings.TrimSpace(msg[idx+2:])
-	if pIdx := strings.Index(m, " ("); pIdx >= 0 {
-		m = m[:pIdx]
-	}
-	return m
-}
-
-func extractCopilotToolName(raw json.RawMessage) string {
-	if len(strings.TrimSpace(string(raw))) == 0 {
-		return ""
-	}
-
-	var tool struct {
-		Name     string `json:"name"`
-		ToolName string `json:"toolName"`
-		Tool     string `json:"tool"`
-	}
-	if err := json.Unmarshal(raw, &tool); err != nil {
-		return ""
-	}
-
-	candidates := []string{tool.Name, tool.ToolName, tool.Tool}
-	for _, candidate := range candidates {
-		candidate = strings.TrimSpace(candidate)
-		if candidate != "" {
-			return candidate
-		}
-	}
-	return ""
-}
-
-func isCopilotMutatingTool(toolName string) bool {
-	name := strings.ToLower(strings.TrimSpace(toolName))
-	if name == "" {
-		return false
-	}
-	return strings.Contains(name, "edit") ||
-		strings.Contains(name, "write") ||
-		strings.Contains(name, "create") ||
-		strings.Contains(name, "delete") ||
-		strings.Contains(name, "rename") ||
-		strings.Contains(name, "move") ||
-		strings.Contains(name, "replace")
-}
-
-func extractCopilotToolCommand(raw json.RawMessage) string {
-	var payload any
-	if json.Unmarshal(raw, &payload) != nil {
-		return ""
-	}
-	var command string
-	var walk func(v any)
-	walk = func(v any) {
-		if command != "" || v == nil {
-			return
-		}
-		switch value := v.(type) {
-		case map[string]any:
-			for key, child := range value {
-				k := strings.ToLower(strings.TrimSpace(key))
-				if k == "command" || k == "cmd" || k == "script" || k == "shell_command" {
-					if s, ok := child.(string); ok {
-						command = strings.TrimSpace(s)
-						return
-					}
-				}
-			}
-			for _, child := range value {
-				walk(child)
-				if command != "" {
-					return
-				}
-			}
-		case []any:
-			for _, child := range value {
-				walk(child)
-				if command != "" {
-					return
-				}
-			}
-		}
-	}
-	walk(payload)
-	return command
-}
-
-func extractCopilotToolPaths(raw json.RawMessage) []string {
-	var payload any
-	if json.Unmarshal(raw, &payload) != nil {
-		return nil
-	}
-
-	pathHints := map[string]bool{
-		"path": true, "paths": true, "file": true, "files": true, "filepath": true, "file_path": true,
-		"cwd": true, "dir": true, "directory": true, "target": true, "pattern": true, "glob": true,
-		"from": true, "to": true, "include": true, "exclude": true,
-	}
-
-	candidates := make(map[string]bool)
-	var walk func(v any, hinted bool)
-	walk = func(v any, hinted bool) {
-		switch value := v.(type) {
-		case map[string]any:
-			for key, child := range value {
-				k := strings.ToLower(strings.TrimSpace(key))
-				childHinted := hinted || pathHints[k] || strings.Contains(k, "path") || strings.Contains(k, "file")
-				walk(child, childHinted)
-			}
-		case []any:
-			for _, child := range value {
-				walk(child, hinted)
-			}
-		case string:
-			if !hinted {
-				return
-			}
-			for _, token := range extractCopilotPathTokens(value) {
-				candidates[token] = true
-			}
-		}
-	}
-	walk(payload, false)
-
-	out := make([]string, 0, len(candidates))
-	for c := range candidates {
-		out = append(out, c)
-	}
-	sort.Strings(out)
-	return out
-}
-
-func extractCopilotPathTokens(raw string) []string {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		return nil
-	}
-	fields := strings.Fields(raw)
-	if len(fields) == 0 {
-		fields = []string{raw}
-	}
-
-	var out []string
-	for _, field := range fields {
-		token := strings.Trim(field, "\"'`()[]{}<>,:;")
-		if token == "" {
-			continue
-		}
-		lower := strings.ToLower(token)
-		if strings.HasPrefix(lower, "http://") || strings.HasPrefix(lower, "https://") || strings.HasPrefix(lower, "file://") {
-			continue
-		}
-		if strings.HasPrefix(token, "-") {
-			continue
-		}
-		if !strings.Contains(token, "/") && !strings.Contains(token, "\\") && !strings.Contains(token, ".") {
-			continue
-		}
-		token = strings.TrimPrefix(token, "./")
-		if token == "" {
-			continue
-		}
-		out = append(out, token)
-	}
-	return lo.Uniq(out)
-}
-
-func estimateCopilotToolLineDelta(raw json.RawMessage) (added int, removed int) {
-	var payload any
-	if json.Unmarshal(raw, &payload) != nil {
-		return 0, 0
-	}
-	lineCount := func(text string) int {
-		text = strings.TrimSpace(text)
-		if text == "" {
-			return 0
-		}
-		return strings.Count(text, "\n") + 1
-	}
-	var walk func(v any)
-	walk = func(v any) {
-		switch value := v.(type) {
-		case map[string]any:
-			var oldText, newText string
-			for _, key := range []string{"old_string", "old_text", "from", "replace"} {
-				if rawValue, ok := value[key]; ok {
-					if s, ok := rawValue.(string); ok {
-						oldText = s
-						break
-					}
-				}
-			}
-			for _, key := range []string{"new_string", "new_text", "to", "with"} {
-				if rawValue, ok := value[key]; ok {
-					if s, ok := rawValue.(string); ok {
-						newText = s
-						break
-					}
-				}
-			}
-			if oldText != "" || newText != "" {
-				removed += lineCount(oldText)
-				added += lineCount(newText)
-			}
-			if rawValue, ok := value["content"]; ok {
-				if s, ok := rawValue.(string); ok {
-					added += lineCount(s)
-				}
-			}
-			for _, child := range value {
-				walk(child)
-			}
-		case []any:
-			for _, child := range value {
-				walk(child)
-			}
-		}
-	}
-	walk(payload)
-	return added, removed
-}
-
-func inferCopilotLanguageFromPath(path string) string {
-	p := strings.ToLower(strings.TrimSpace(path))
-	if p == "" {
-		return ""
-	}
-	base := strings.ToLower(filepath.Base(p))
-	switch base {
-	case "dockerfile":
-		return "docker"
-	case "makefile":
-		return "make"
-	}
-	switch strings.ToLower(filepath.Ext(p)) {
-	case ".go":
-		return "go"
-	case ".py":
-		return "python"
-	case ".ts", ".tsx":
-		return "typescript"
-	case ".js", ".jsx":
-		return "javascript"
-	case ".tf", ".tfvars", ".hcl":
-		return "terraform"
-	case ".sh", ".bash", ".zsh", ".fish":
-		return "shell"
-	case ".md", ".mdx":
-		return "markdown"
-	case ".json":
-		return "json"
-	case ".yml", ".yaml":
-		return "yaml"
-	case ".sql":
-		return "sql"
-	case ".rs":
-		return "rust"
-	case ".java":
-		return "java"
-	case ".c", ".h":
-		return "c"
-	case ".cc", ".cpp", ".cxx", ".hpp":
-		return "cpp"
-	case ".rb":
-		return "ruby"
-	case ".php":
-		return "php"
-	case ".swift":
-		return "swift"
-	case ".vue":
-		return "vue"
-	case ".svelte":
-		return "svelte"
-	case ".toml":
-		return "toml"
-	case ".xml":
-		return "xml"
-	}
-	return ""
-}
-
-func formatModelMap(m map[string]int, unit string) string {
-	if len(m) == 0 {
-		return ""
-	}
-	parts := make([]string, 0, len(m))
-	for model, count := range m {
-		parts = append(parts, fmt.Sprintf("%s: %d %s", model, count, unit))
-	}
-	sort.Strings(parts)
-	return strings.Join(parts, ", ")
-}
-
-func formatModelMapPlain(m map[string]int) string {
-	if len(m) == 0 {
-		return ""
-	}
-	parts := make([]string, 0, len(m))
-	for model, count := range m {
-		parts = append(parts, fmt.Sprintf("%s: %d", model, count))
-	}
-	sort.Strings(parts)
-	return strings.Join(parts, ", ")
-}
-
-func setRawInt(snap *core.UsageSnapshot, key string, v int) {
-	if v > 0 {
-		snap.Raw[key] = fmt.Sprintf("%d", v)
-	}
-}
-
-func setRawStr(snap *core.UsageSnapshot, key, v string) {
-	if v != "" {
-		snap.Raw[key] = v
-	}
-}
-
 func firstNonNilFloat(values ...*float64) *float64 {
 	for _, v := range values {
 		if v != nil {
@@ -1978,34 +479,4 @@ func clampPercent(v float64) float64 {
 		return 100
 	}
 	return v
-}
-
-func sanitizeMetricName(name string) string {
-	name = strings.ToLower(strings.TrimSpace(name))
-	if name == "" {
-		return "unknown"
-	}
-
-	var b strings.Builder
-	lastUnderscore := false
-	for _, r := range name {
-		switch {
-		case r >= 'a' && r <= 'z':
-			b.WriteRune(r)
-			lastUnderscore = false
-		case r >= '0' && r <= '9':
-			b.WriteRune(r)
-			lastUnderscore = false
-		default:
-			if !lastUnderscore {
-				b.WriteByte('_')
-				lastUnderscore = true
-			}
-		}
-	}
-	out := strings.Trim(b.String(), "_")
-	if out == "" {
-		return "unknown"
-	}
-	return out
 }
